requirements.txt
streamlit
openai
python-docx
PyPDF2
markdown